src/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.18
conekta==2.3.4
coreapi==2.3.0
coreschema==0.0.4
decorator==4.0.11
Django==1.9.13
django-common-helpers==0.9.1
django-cors-headers==1.1.0
django-cron==0.5.0
django-debug-toolbar==1.6
django-extensions==1.6.7
django-filter==0.11.0
django-haystack==2.5.1
django-rest-swagger==0.3.10
djangorestframework==3.3.3
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.7.2
drf-haystack==1.6.1
drf-nested-routers==0.10.0
enum34==1.1.6
httplib2==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==1.0
nose==1.3.7
openapi-codec==1.3.1
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.1.2
prompt-toolkit==1.0.14
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
pysolr==3.5.0
python-dateutil==2.6.0
python-memcached==1.53
pytz==2017.2
PyYAML==3.12
requests==2.14.2
scandir==1.5
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
Werkzeug==0.12.1
